requirements.txt
streamlit>=1.28.0
pandas>=2.1.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
tensorflow>=2.14.0
torch>=2.2.2
faiss-cpu>=1.7.4
pypdf>=3.17.4
joblib>=1.3.0
requests>=2.31.0
huggingface-hub>=0.16.4
transformers==4.39.3
